+++ sources/packages/backend/apps/api/src/app.system-access.module.ts
@@ -16,8 +16,6 @@ import {
   SFASPartTimeApplicationsService,
   StudentFileService,
   MSFAANumberService,
-  CRAPersonalVerificationService,
-  CRAIntegrationService,
   SFASIndividualService,
   SFASRestrictionService,
   DisbursementScheduleErrorsService,
@@ -35,7 +33,6 @@ import {
 import {
   AssessmentControllerService,
   ApplicationSystemAccessController,
-  CRAIntegrationSystemAccessController,
   MSFAAIntegrationSystemAccessController,
   SFASIntegrationSystemAccessController,
   ECertIntegrationSystemAccessController,
@@ -82,7 +79,6 @@ import { TypeOrmModule } from "@nestjs/typeorm";
     ATBCSystemAccessController,
     SINValidationSystemAccessController,
     ApplicationSystemAccessController,
-    CRAIntegrationSystemAccessController,
     MSFAAIntegrationSystemAccessController,
     SFASIntegrationSystemAccessController,
     ECertIntegrationSystemAccessController,
@@ -107,8 +103,6 @@ import { TypeOrmModule } from "@nestjs/typeorm";
     SFASPartTimeApplicationsService,
     StudentFileService,
     MSFAANumberService,
-    CRAPersonalVerificationService,
-    CRAIntegrationService,
     SshService,
     MSFAARequestService,
     MSFAAIntegrationService,
