+++ sources/packages/backend/apps/api/src/route-controllers/application/application.system.controller.e2e-spec.ts
@@ -35,7 +35,6 @@ import {
   createFakeEducationProgramOffering,
 } from "../../testHelpers/fake-entities";
 import { createMockedJwtService } from "../../testHelpers/mocked-providers/jwt-service-mock";
-import { CraIntegrationModule } from "../../cra-integration/cra-integration.module";
 import { ApplicationSystemAccessController } from "./application.system-access.controller";
 import { createFakeInstitution, createFakeUser } from "@sims/test-utils";
 import { ConfigModule } from "@sims/utilities/config";
@@ -58,7 +57,7 @@ describe.skip("Test system-access/application Controller", () => {
     );
     accesstoken = token.access_token;
     const module: TestingModule = await Test.createTestingModule({
-      imports: [DatabaseModule, AuthModule, CraIntegrationModule, ConfigModule],
+      imports: [DatabaseModule, AuthModule, ConfigModule],
       controllers: [ApplicationSystemAccessController],
       providers: [
         ApplicationService,
