+++ sources/packages/backend/apps/api/src/route-controllers/index.ts
@@ -2,7 +2,6 @@ export * from "./user/user.controller";
 export * from "./program-year/program-year.controller";
 export * from "./config/config.controller";
 export * from "./dynamic-form/dynamic-form.controller";
-export * from "./cra-integration/cra-integration.system-access.controller";
 export * from "./education-program/education-program.controller.service";
 export * from "./education-program/education-program.aest.controller";
 export * from "./education-program/education-program.institutions.controller";
