+++ sources/packages/backend/apps/api/src/services/index.ts
@@ -10,9 +10,7 @@ export * from "./form/form.service";
 export * from "./form/constants";
 export * from "./application/application.service";
 export * from "./institution-location/institution-location.service";
-export * from "../cra-integration/cra-integration.service";
 export * from "./cra-income-verification/cra-income-verification.service";
-export * from "../cra-integration/cra-personal-verification.service";
 export * from "./institution-user-auth/institution-user-auth.service";
 export * from "./education-program-offering/education-program-offering.service";
 export * from "./education-program/education-program.service";
