+++ sources/packages/backend/apps/queue-consumers/src/processors/schedulers/cra-integration/models/cra-validation-result.dto.ts
@@ -1,4 +1,4 @@
-export class CRAValidationResultAPIOutDTO {
+export class CRAValidationResultQueueOutDTO {
   generatedFile: string;
   uploadedRecords: number;
 }
