+++ sources/packages/backend/apps/queue-consumers/src/processors/schedulers/cra-integration/models/process-response.dto.ts
@@ -1,4 +1,4 @@
-export class ProcessResponseAPIOutDTO {
+export class ProcessResponseQueueOutDTO {
   processSummary: string[];
   errorsSummary: string[];
 }
