+++ sources/packages/backend/libs/integrations/src/cra-integration/__mocks__/SshServiceMock.ts
