+++ sources/packages/backend/libs/integrations/src/cra-integration/__mocks__/response-folder/CCRA_RESPONSE_ABCSL00001.TXT
