+++ sources/packages/backend/libs/integrations/src/cra-integration/__mocks__/response-folder/CCRA_RESPONSE_ABCSL00002.TXT
