+++ sources/packages/backend/libs/integrations/src/cra-integration/cra-files/cra-file-footer.ts
