+++ sources/packages/backend/libs/integrations/src/cra-integration/cra-files/cra-response-total-income-record.ts
