+++ sources/packages/backend/libs/integrations/src/services/index.ts
@@ -1,3 +1,4 @@
 export * from "./ssh/ssh.service";
 export * from "./ssh/sftp-integration-base";
 export * from "./student-assessment/student-assessment.service";
+export * from "./cra-income-verification/cra-income-verification.service";
